gcp/cloudrun/bookstore-server/requirements.txt
grpcio==1.37.1
protobuf==3.17.0
google-auth==1.30.0
six==1.16.0
googleapis-common-protos==1.53.0
